pjt_server/requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.3
django-allauth==0.43.0
django-appconf==1.0.4
django-cors-headers==3.5.0
django-imagekit==4.0.2
django-rest-auth==0.9.5
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
pilkit==2.0
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
